cleanup assets code structure
