+++ apps/designer/app/designer/features/sidebar-left/panels/asset-manager/uploading-animation.tsx
@@ -3,36 +3,21 @@ import { ProgressRadial } from "@webstudio-is/design-system";
 import { useState } from "react";
 import { useInterval } from "react-use";
 
-const useFakeProgress = ({
-  isUploading,
-  isDeleting,
-}: {
-  isUploading: boolean;
-  isDeleting: boolean;
-}) => {
+const useFakeProgress = () => {
   const [progressBarPercentage, setProgressBarPercentage] = useState(0);
 
   // @todo rewrite this fake indication to show real progress
-  useInterval(
-    () => {
-      setProgressBarPercentage((percentage) =>
-        percentage < 60 ? percentage + 1 : percentage
-      );
-    },
-    isUploading || isDeleting ? 100 : null
-  );
+  useInterval(() => {
+    setProgressBarPercentage((percentage) =>
+      percentage < 60 ? percentage + 1 : percentage
+    );
+  }, 100);
 
   return progressBarPercentage;
 };
 
-export const UploadingAnimation = ({
-  isUploading,
-  isDeleting,
-}: {
-  isUploading: boolean;
-  isDeleting: boolean;
-}) => {
-  const progressBarPercentage = useFakeProgress({ isUploading, isDeleting });
+export const UploadingAnimation = () => {
+  const progressBarPercentage = useFakeProgress();
   return (
     <ProgressRadial
       css={{ position: "relative", zIndex: 10 }}
